refactor: move thread runtime stream loop
